requirements.txt
django==5.1.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg2==2.9.10
environs==11.0.0
